+++ src/structures/GpuBruteForceNN.h
@@ -14,10 +14,10 @@ class GpuBruteForceNN : public INearestNeighbor<4>
 	/// \param points
 	void Build(const std::vector<point4_t>& points) override;
 
-	bool FindNearestNeighbor(const point4_t& point, size_t& index) override;
+	bool FindNearestNeighbor(const point4_t& point, uint32_t& index) override;
 	bool FindNearestNeighbors(const point4_t& point,
-		float radius,
-		std::vector<size_t>& indices) override;
+	                          float radius,
+	                          std::vector<uint32_t>& indices) override;
  private:
 	OpenCLContext& m_openCLContext;
 	std::unique_ptr<cl::Buffer> m_pointBuffer;
